sfsf

+++ src/com/joogle/model/TermWithWeight.java
@@ -2,20 +2,20 @@ package com.joogle.model;
 
 public class TermWithWeight implements Comparable<TermWithWeight> {
 	public String term;
-	public Integer freq;
+	public Double freq;
 
 	public TermWithWeight() {
 
 	}
 
-	public TermWithWeight(String term, Integer freq) {
+	public TermWithWeight(String term, Double freq) {
 		this.term = term;
 		this.freq = freq;
 	}
 
 	@Override
 	public int compareTo(TermWithWeight t) {
-		return (t.freq - this.freq);
+		return (int)(t.freq - this.freq);
 	}
 
 	public String toString() {
